refactor(version): extract build metadata into internal/version
Each binary previously had its own Version/Commit/BuildTime vars and
duplicated getVersion() helper. Lifts the three vars plus the "" → "dev"
fallback into a shared internal/version package so:

- The bls binary (and future uki binary) reuse the same metadata source
  instead of redefining it per-main.
- The ldflags target moves back to the full-import-path form
  (github.com/.../internal/version.Version), which is identical for every
  binary built from this repo — release scripts don't need per-binary
  variants.

CHANGES
- internal/version: new package with Version/Commit/BuildTime vars and
  String() helper. TestString moved here from cmd/refind-btrfs-snapshots
  (it was always a package-level test, never binary-specific).
- cmd/refind-btrfs-snapshots/root.go: drop local vars + getVersion(),
  read via version.String() / version.Commit / version.BuildTime.
- cmd/refind-btrfs-snapshots/version.go: same swap.
- cmd/refind-btrfs-snapshots/{root,version}_test.go: tests retarget the
  shared package's vars.
- build-release.yml + docs/USAGE.md: ldflags use the internal/version
  import path.

VERIFIED
- go test ./... — all packages pass including new internal/version.
- Local ldflags injection at -X internal/version.Version=injected-1.2.3
  produced the expected version output.
