+++ .github/workflows/build-release.yml
@@ -144,9 +144,9 @@ jobs:
           DATE="$(date -u +%Y-%m-%dT%H:%M:%SZ)"
 
           go build -ldflags "-s -w \
-            -X 'main.Version=${VERSION}' \
-            -X 'main.Commit=${COMMIT}' \
-            -X 'main.BuildTime=${DATE}'" \
+            -X 'github.com/jmylchreest/refind-btrfs-snapshots/internal/version.Version=${VERSION}' \
+            -X 'github.com/jmylchreest/refind-btrfs-snapshots/internal/version.Commit=${COMMIT}' \
+            -X 'github.com/jmylchreest/refind-btrfs-snapshots/internal/version.BuildTime=${DATE}'" \
             -o "dist/refind-btrfs-snapshots" ./cmd/refind-btrfs-snapshots
 
       - name: Prepare distribution
